Use 32 bits for CCV ID
Some datasets executed with more than 2^16 decision boundaries

+++ src/core/algorithms/md/hymd/column_classifier_value_id.h
@@ -3,5 +3,5 @@
 #include <cstdint>
 
 namespace algos::hymd {
-using ColumnClassifierValueId = std::uint16_t;
+using ColumnClassifierValueId = std::uint32_t;
 }  // namespace algos::hymd
